height and pointer changes
I made a tree pointer for the double tree pointer to point to because
otherwise it didn’t work with my compiler? Something to do with it not
wanting the pointer inside the double pointer to be NULL.
I also wrote the getHeight() function which I believe works and added
some print statements to the traverse() function that show which nodes
are to the left and which nodes are to the right; uncomment these if
desired to help debug the tree.